Move JsObjectId to js_collections module
